requirements.txt
fastapi
uvicorn
openai
PyMuPDF
python-docx
python-multipart